refactor: introduce a JsPackageManager abstract class and specialized subclasses for NPM, Yarn and Yarn 2.
The goal of this will be to use the Factory Pattern to encapsulate everything specific to a package manager in a single class.
Then a factory will return an instance of NPM, Yarn or Yarn 2 proxy class according to the package manager used/prefered in end project.
All CLI code will be package manager agnostic and will just call functions of JsPackageManager interface.